Fixed gcmlcm() function's ecdf first element pop()
An inconsistency was left out after the ecdf function was fixed to output only the unique x values.
The gcmlcm() function was popping the first (x_0, 0.0) element of the ecdf in the prior incorrect ecdf definition. 
This caused the specific case where the dataset was linear in all places except for the first point to pass through the KS test but then become linear in the gcmlcm() function, causing qhull to break.
The pop() lines have been removed as they are no longer relevant.